Added advantage 1 when dual-wielding
Currently you need feats to get anything out of dual-wielding 2 one-handed weapons.
When choosing to wield 2 weapons, you are choosing to not carry a shield. So there should be some kind of inherent benefit to it. Attacking with 2 one-handed weapons should not be the same as attacking with a single one-handed weapon while wearing a shield in the other hand.
And attacking with 2 one-handed weapons is comparable to attacking with a single two-handed weapon. In the sense that you use both hands for attacking. And don't have a free hand for a shield.

Another thing: If you take the Multi-Attack Specialist feat to simulate attacking with both weapons, you would actually have an easier time multi-attacking with a single two-handed weapon, because of the extra advantage 1 you get, just for it being a two-handed weapon. That seems wrong. Multi-attacking should not be easier with a two-handed weapon than when dual-wielding 2 one-handed weapons.